# Bounty Solution (2.4x speedup)