--- a/fastgs/optimizer/include/adam.h
+++ b/fastgs/optimizer/include/adam.h
@@ -1,0 +1,18 @@
+#pragma once
+
+namespace fast_gs::optimizer {
+
+    void adam_step(
+        float* param,
+        float* exp_avg,
+        float* exp_avg_sq,
+        const float* param_grad,
+        const int n_elements,
+        const float lr,
+        const float beta1,
+        const float beta2,
+        const float eps,
+        const float bias_correction1_rcp,
+        const float bias_correction2_sqrt_rcp);
+
+}