… CMakeLists.txt
… README.md
… fastgs/CMakeLists.txt
… fastgs/rasterization/README.md
… fastgs/rasterization/include/kernel_utils.cuh
… fastgs/rasterization/include/kernels_backward.cuh
… fastgs/rasterization/include/kernels_forward.cuh
… fastgs/rasterization/src/forward.cu
… fastgs/rasterization/src/rasterization_api.cu
… fastgs/utils/helper_math.h
… include/core/dataset.hpp
… include/core/fast_rasterizer.hpp
… include/core/fast_rasterizer_autograd.hpp
… include/core/fused_adam.hpp
… include/core/image_io.hpp
… include/core/mcmc.hpp
… include/core/splat_data.hpp
… src/camera.cpp
… src/fast_rasterizer.cpp
… src/fast_rasterizer_autograd.cpp
… src/fused_adam.cpp
… src/image_io.cpp
… src/main.cpp
… src/mcmc.cpp
… src/splat_data.cpp
… src/trainer.cpp