@MrNeRF 2025-08-03
Final eval:
| scene   |   iteration |    psnr |     ssim |    lpips | num_gaussians   |
|:--------|------------:|--------:|---------:|---------:|:----------------|
| garden  |       30000 | 27.2306 | 0.858708 | 0.156047 | 1,000,000       |
| bicycle |       30000 | 25.4351 | 0.777493 | 0.255592 | 1,000,000       |
| stump   |       30000 | 26.7973 | 0.790044 | 0.263982 | 1,000,000       |
| bonsai  |       30000 | 32.6185 | 0.948317 | 0.249532 | 1,000,000       |
| counter |       30000 | 29.2851 | 0.917415 | 0.241510 | 1,000,000       |
| kitchen |       30000 | 31.3604 | 0.933028 | 0.155878 | 1,000,000       |
| room    |       30000 | 32.2236 | 0.931001 | 0.273806 | 1,000,000       |
| mean    |       30000 | 29.2786 | 0.879429 | 0.228050 | 1,000,000       |