@fhahlbohm 2025-08-04
Thanks for the kind words!

> I'm curious about the individual contribution of the New Rasterizer to the overall performance gain. Since you implemented multiple optimizations (New Rasterizer, MCMC, fused Adam, data loading), do you happen to know roughly how much of the speedup came specifically from the rasterizer changes versus the other improvements?

Yes, a small write-up with a bunch of experiments beyond the scope of this bounty is coming soon. Together with a PyTorch API for the rasterizer for the PyTorch-based research community.