@andorxornot 2025-08-04
Wow, that's incredibly cool. Quite a lot of changes and most of them are pure optimizations. The papers that will give even more speed have not yet been implemented. Crazy!